removed inline editing when there is formula for timeline and richtext
