+++ scripts/scripts.js
@@ -240,3 +240,29 @@ function handleMathJax(){
 window.onload = ()=>{
   handleMathJax();
 }
+
+export function removeDataAueAttributesWhenThereIsFormula(element) {
+  if (!element || !(element instanceof HTMLElement)) {
+      console.error("O argumento fornecido não é um elemento HTML válido.");
+      return;
+  }
+
+  if(!isInEditor) return;
+
+  //Se não houver fórmula, morre aqui
+  const body = element.textContent;
+  if(!body.match(/(?:\$|\\\(|\\\[|\\begin\{.*?})/)) return;
+
+  // Seleciona todos os elementos filhos do elemento fornecido
+  const children = element.querySelectorAll("*");
+
+  children.forEach(child => {
+      // Itera sobre os atributos do elemento filho
+      Array.from(child.attributes).forEach(attr => {
+          // Verifica se o nome do atributo começa com "data-aue"
+          if (attr.name.startsWith("data-aue")) {
+              child.removeAttribute(attr.name); // Remove o atributo
+          }
+      });
+  });
+}
